+++ plugins/todo-backend/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-todo-backend
 
+## 0.1.33-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/catalog-client@1.1.0-next.2
+  - @backstage/catalog-model@1.1.1-next.0
+  - @backstage/config@1.0.2-next.0
+  - @backstage/errors@1.1.1-next.0
+  - @backstage/integration@1.3.1-next.2
+  - @backstage/backend-common@0.15.1-next.3
+
 ## 0.1.33-next.1
 
 ### Patch Changes